+++ src/components/App.js
@@ -1,8 +1,7 @@
 import React from "react";
 import FormComponent from "./YAMLComponents/FormComponent";
 import EndPage from "./frontPageComponents/EndPage";
-import MarkdownEditor from "./mdPreviewComponents/MarkdownEditor";
-// import Editor from "./mdPreviewComponents/Editor";
+import Editor from "./mdPreviewComponents/editor/Editor";
 import IndexPage from "./frontPageComponents/IndexPage";
 import Overview from "./mypage/Overview";
 import { BrowserRouter, Route } from "react-router-dom";
@@ -15,7 +14,7 @@ const App = () => {
           <Route exact path="/" component={IndexPage} />
           <Route path="/myPage" component={Overview} />
           <Route path="/createNewLesson" component={FormComponent} />
-          <Route path="/editor" component={MarkdownEditor} />
+          <Route path="/editor" component={Editor} />
           <Route exact path="/endpage" component={EndPage} />
         </div>
       </BrowserRouter>
