+++ src/components/YAMLComponents/FormComponent.js
@@ -1,11 +1,9 @@
 import "./formpage.css";
 import React from "react";
 import { connect } from "react-redux";
-import { Redirect } from "react-router-dom";
 import FormPage from "./FormPage";
 import COURSELIST from "./settingsFiles/COURSELIST";
 import { LANGUAGES } from "./settingsFiles/languages/formpage_NO";
-import GoogleAuth from "../GoogleAuth";
 
 class FormComponent extends React.Component {
   constructor(props) {
