+++ src/components/YAMLComponents/Page3.js
@@ -98,14 +98,14 @@ const Page3 = props => {
           </div>
         </div>
       </div>
-      <div style={{ margin: "30px" }} />
+
       <div id="smallScreen" className="ui grid">
         <div className="sixteen wide column">
           <div className="field">
             <label>
               <h3>{YML_TEXT.subject}</h3>
             </label>
-            <div style={{ marginRight: "-30px" }}>
+            <div>
               <Dropdown
                 placeholder="Velg fag"
                 fluid
@@ -118,13 +118,13 @@ const Page3 = props => {
             </div>
           </div>
         </div>
-        <div style={{ margin: "30px" }} />
+
         <div className="sixteen wide column">
           <div className="field">
             <label>
               <h3>{YML_TEXT.topic}</h3>
             </label>
-            <div style={{ marginLeft: "-13px", marginRight: "-15px" }}>
+            <div>
               <Dropdown
                 placeholder="Velg tema"
                 fluid
@@ -137,7 +137,6 @@ const Page3 = props => {
             </div>
           </div>
         </div>
-        <div style={{ margin: "100px" }} />
       </div>
     </div>
   );
