+++ src/components/YAMLComponents/formpage.css
@@ -11,7 +11,4 @@
   #bigScreen {
     display: none;
   }
-  .ui.form {
-    height: 600px;
-  }
 }
