--- src/components/mdPreviewComponents/largeScreen/ControlPanel.js
@@ -1,182 +1,0 @@
-import React from "react";
-import { connect } from "react-redux";
-import Buttons from "./Button";
-import {
-  emphasis,
-  undoRedo,
-  saveLoadNew,
-  image,
-  lists,
-  sections,
-  code
-} from "../settingsFiles/buttonConfig.js";
-
-class ControlPanel extends React.Component {
-  handleButtonClick = (
-    bTitle,
-    output,
-    cursorIntON,
-    cursorIntOFF,
-    endOutput
-  ) => {
-    this.props.handleButtonClick(
-      bTitle,
-      output,
-      cursorIntON,
-      cursorIntOFF,
-      endOutput
-    );
-  };
-
-  render() {
-    return (
-      <div className="ui segment grid buttonBorder">
-        <div className="ui icon buttons">
-          {emphasis.map(element => (
-            <div key={element.bTitle} className="">
-              <Buttons
-                bTitle={element.bTitle}
-                icon={element.icon}
-                output={element.output}
-                title={element.title}
-                cursorIntON={element.cursorIntON}
-                cursorIntOFF={element.cursorIntOFF}
-                endOutput={element.endOutput}
-                onButtonClick={this.handleButtonClick}
-                shortcutKey={element.shortcut}
-              />
-            </div>
-          ))}
-        </div>
-        <div className="space" />
-        <div className="ui icon buttons">
-          {undoRedo.map(element => (
-            <div key={element.bTitle} className="">
-              <Buttons
-                bTitle={element.bTitle}
-                icon={element.icon}
-                output={element.output}
-                title={element.title}
-                cursorIntON={element.cursorIntON}
-                cursorIntOFF={element.cursorIntOFF}
-                endOutput={element.endOutput}
-                onButtonClick={this.handleButtonClick}
-                shortcutKey={element.shortcut}
-              />
-            </div>
-          ))}
-        </div>
-        <div className="space" />
-        <div className="ui icon buttons">
-          {saveLoadNew.map(element => (
-            <div key={element.bTitle} className="">
-              <Buttons
-                bTitle={element.bTitle}
-                icon={element.icon}
-                output={element.output}
-                title={element.title}
-                cursorIntON={element.cursorIntON}
-                cursorIntOFF={element.cursorIntOFF}
-                endOutput={element.endOutput}
-                onButtonClick={this.handleButtonClick}
-                shortcutKey={element.shortcut}
-              />
-            </div>
-          ))}
-        </div>
-        <div className="space" />
-        <div className="ui icon buttons">
-          {image.map(element => (
-            <div key={element.bTitle} className="">
-              <Buttons
-                bTitle={element.bTitle}
-                icon={element.icon}
-                output={element.output}
-                title={element.title}
-                cursorIntON={element.cursorIntON}
-                cursorIntOFF={element.cursorIntOFF}
-                endOutput={element.endOutput}
-                onButtonClick={this.handleButtonClick}
-                shortcutKey={element.shortcut}
-              />
-            </div>
-          ))}
-        </div>
-        <div className="space" />
-        <div className="ui icon buttons">
-          {lists.map(element => (
-            <div key={element.bTitle} className="">
-              <Buttons
-                bTitle={element.bTitle}
-                icon={element.icon}
-                output={element.output}
-                title={element.title}
-                cursorIntON={element.cursorIntON}
-                cursorIntOFF={element.cursorIntOFF}
-                endOutput={element.endOutput}
-                onButtonClick={this.handleButtonClick}
-                shortcutKey={element.shortcut}
-              />
-            </div>
-          ))}
-        </div>
-        <div className="space" />
-        <div className="ui icon buttons">
-          {sections.map(element => (
-            <div key={element.bTitle} className="sections">
-              <Buttons
-                bTitle={element.bTitle}
-                icon={element.icon}
-                output={element.output}
-                title={element.title}
-                cursorIntON={element.cursorIntON}
-                cursorIntOFF={element.cursorIntOFF}
-                endOutput={element.endOutput}
-                onButtonClick={this.handleButtonClick}
-                shortcutKey={element.shortcut}
-              />
-            </div>
-          ))}
-        </div>
-        <div className="space" />
-        <div className="ui icon buttons">
-          {code.map(element => (
-            <div key={element.bTitle} className="sections">
-              <Buttons
-                bTitle={element.bTitle}
-                icon={element.icon}
-                output={element.output}
-                title={element.title}
-                cursorIntON={element.cursorIntON}
-                cursorIntOFF={element.cursorIntOFF}
-                endOutput={element.endOutput}
-                onButtonClick={this.handleButtonClick}
-                shortcutKey={element.shortcut}
-              />
-            </div>
-          ))}
-        </div>
-        <div className="right floated two wide column">
-          <div
-            style={{ marginTop: "-50px", marginRight: "-1000px" }}
-            className="ui avatar image"
-          >
-            {this.props.imageUrl ? (
-              <img src={this.props.imageUrl} alt="useImage"></img>
-            ) : (
-              <i className="user icon"></i>
-            )}
-          </div>
-        </div>
-      </div>
-    );
-  }
-}
-
-const mapStateToProps = state => {
-  return {
-    imageUrl: state.auth.imageUrl
-  };
-};
-
-export default connect(mapStateToProps)(ControlPanel);
