+++ src/components/mypage/LeftBox.js
@@ -1,18 +1,15 @@
-import React from 'react';
+import React from "react";
 import "./LeftBoxStyle.css";
 
-function LeftBox(){
+function LeftBox() {
   return (
-    <div className = "LeftBox">
-      <div className = "informationTable">
-      <p id="header">Hopp til</p>
+    <div className="LeftBox">
+      <div className="informationTable">
+        <p id="header">Hopp til</p>
       </div>
       <p id="text1">Python</p>
       <p id="text2">JavaScript</p>
-
     </div>
-
-  )
-
+  );
 }
 export default LeftBox;
