+++ src/index.js
@@ -3,7 +3,7 @@ import "semantic-ui-css/semantic.min.css";
 import React from "react";
 import ReactDOM from "react-dom";
 import { Provider } from "react-redux";
-import { createStore, applyMiddleware } from "redux";
+import { createStore } from "redux";
 
 import App from "./components/App";
 import reducers from "./reducers";
